Extensions for GPUs
