+++ Project.toml
@@ -4,14 +4,20 @@ authors = ["Alexis Montoison <alexis.montoison@polymtl.ca>"]
 version = "0.1.0"
 
 [deps]
-AMDGPU = "21141c5a-9bdb-4563-92ae-f87d6854732e"
 Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
-CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 KernelAbstractions = "63c18a36-062a-441e-b654-da1e3ab1ce7c"
 LightGraphs = "093fc24a-ae57-5d10-9952-331d41423f4d"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Metis = "2679e427-3c69-5b7f-982b-ece356f1e94b"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
+
+[weakdeps]
+AMDGPU = "21141c5a-9bdb-4563-92ae-f87d6854732e"
+CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
+
+[extensions]
+KrylovPreconditionersAMDGPUExt = "AMDGPU"
+KrylovPreconditionersCUDAExt = "CUDA"
 
 [compat]
 AMDGPU = "0.7"
@@ -21,10 +27,3 @@ Metis = "1"
 LightGraphs = "1"
 KernelAbstractions = "0.9"
 julia = "1.9"
-
-[extras]
-Krylov = "ba0b0d4f-ebba-5204-a429-3ac8c609bfb7"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[targets]
-test = ["Krylov", "Test"]
